# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для хранения и менеджмента пользовательских данных",
        "title": "GophKeeper API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Создаёт нового пользователя и возвращает JWT токены",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {}
                    },
                    "409": {
                        "description": "Пользователь уже существует",
                        "schema": {}
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {}
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создаёт нового пользователя и возвращает JWT токены",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Register.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная регистрация"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Пользователь уже существует",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/account": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый аккаунт",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Создать новый аккаунт",
                "parameters": [
                    {
                        "description": "Данные для сохранения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Account.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Конфликт создания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/accounts/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Обновить существующий аккаунт",
                "parameters": [
                    {
                        "description": "Данные для обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Account.PutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет аккаунт с входящим ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Удалить существующий аккаунт",
                "parameters": [
                    {
                        "description": "Данные для с id записью",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Account.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/accounts/fetch": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список данных аккаунтов пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Получить все аккаунты",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Account.DeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "Account.PostRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user"
                },
                "metadata": {
                    "type": "string",
                    "example": "metadata"
                },
                "password": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "Account.PutRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "login": {
                    "type": "string",
                    "example": "user"
                },
                "metadata": {
                    "type": "string",
                    "example": "metadata"
                },
                "password": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "Login.PostRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user"
                },
                "password": {
                    "type": "string",
                    "example": "pass"
                }
            }
        },
        "Register.PostRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user"
                },
                "password": {
                    "type": "string",
                    "example": "pass"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Введите \"Bearer \u003cJWT\u003e\", где JWT - ваш access токен",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}